add quotes to args for win32
Apparently, when IPC::Open3 is given a list of command line arguments on
win32, it does not quote strings with spaces in them as it does on sane
systems. This change adds quotes around the argument to perl -e in
t/02-send-recv.t in an attempt to work around this.